Building's area refactored to intiger type.

+++ sv2-2021-jvjbf-masodik-reszvizsga/src/main/java/city/Building.java
@@ -1,27 +1,27 @@
 package city;
 
 public abstract class Building {
-    private double area;
+    private int area;
     private int levels;
     private Address address;
 
-    public Building(double area, int levels, Address address) {
+    public Building(int area, int levels, Address address) {
         this.area = area;
         this.levels = levels;
         this.address = address;
     }
 
-    public Building(double area, Address address) {
+    public Building(int area, Address address) {
         this(area, 1, address);
     }
 
     public abstract int calculateNumberOfPeopleCanFit();
 
-    public double getFullArea() {
+    public int getFullArea() {
         return area * levels;
     }
 
-    public double getArea() {
+    public int getArea() {
         return area;
     }
