+++ sv2-2021-jvjbf-masodik-reszvizsga/src/main/java/city/Home.java
@@ -3,19 +3,19 @@ package city;
 public class Home extends Building {
     public static final int SQUARE_METERS_NEED_PER_PERSON = 20;
 
-    public Home(double area, int levels, Address address) {
+    public Home(int area, int levels, Address address) {
         super(area, levels, address);
         if (levels > 3) {
             throw new IllegalArgumentException("Too high to be home!");
         }
     }
 
-    public Home(double area, Address address) {
+    public Home(int area, Address address) {
         super(area, address);
     }
 
     @Override
     public int calculateNumberOfPeopleCanFit() {
-        return (int) (getFullArea() / SQUARE_METERS_NEED_PER_PERSON);
+        return getFullArea() / SQUARE_METERS_NEED_PER_PERSON;
     }
 }
